Increasing the "learning rate", epsilon
The learning rate, or step size, appears to be too low. Updates to the parameters produces no change in the total dust mass.
Or maybe the way of choosing initial conditions is already quite clever? But still, it's odd that the process only goes through a single iteration before claiming it has completed successfully...